Update bootstrap method
We don't really need a separate exception if the hooking class has not hook method. It's extremely unlikely and just adds more boilerplate
Set the hook prefix as a constant on the core theme class and then retrieve it from the hooking class in theme modules. SPOF
Don't underscore the hooker property: it's not private
Remove unused $m property from classes